requirements.txt
tqdm
langchain
tiktoken
weaviate-client
openai
gradio
python-dotenv
pydantic
beautifulsoup4